Rename log in file structure
