+++ walk-in-interview/src/main/webapp/log-in/applicant-log-in/index.html
@@ -28,5 +28,5 @@
     </div>
   </body>
 
-  <script type="module" src="./applicant-log-in.js"></script>
+  <script type="module" src="./script.js"></script>
 </html>
