+++ walk-in-interview/src/main/webapp/log-in/applicant-log-in/script.js
@@ -1,5 +1,5 @@
 /**
- * This file is specific to log-in/log-in.html. 
+ * This file is specific to log-in/applicant-log-in/index.html. 
  * It renders the fields on the page dynamically
  */
 
@@ -10,11 +10,11 @@ const CurrentLocale = 'en';
  * Import statements are static so its parameters cannot be dynamic.
  * TODO(issue/22): figure out how to use dynamic imports
  */
-import {AppStrings} from '../strings.en.js';
+import {AppStrings} from '../../strings.en.js';
 
 const COMMONG_STRINGS = AppStrings['log-in'];
 const STRINGS = AppStrings['applicant-log-in'];
-const HOMEPAGE_PATH = './log-in.html';
+const HOMEPAGE_PATH = '../index.html';
 
 window.onload = () => {
   renderPageElements();
