+++ walk-in-interview/src/main/webapp/log-in/business-log-in/index.html
@@ -26,5 +26,5 @@
     </div>
   </body>
 
-  <script type="module" src="./business-log-in.js"></script>
+  <script type="module" src="./script.js"></script>
 </html>
